chore: :arrow_down: chore(deps): Update dependencies and fix eslint
Description:
This commit updates several dependencies in the project and fixes eslint configuration.

Details:
The dependencies updated include "lucide-react" from version 0.379.0 to 0.381.0, "swiper" from 11.1.3 to 11.1.4, and "@types/node" from 20.12.12 to 20.13.0. The eslint version has also been updated to 8.57.0 and nodemon to 3.1.2.

Changeset:
The changes were made in the package.json and package-lock.json files. The versions of the dependencies were updated to their latest versions. The eslint configuration was also updated to work with the latest version.

Modified files:
- package.json
- package-lock.json

This commit ensures that the project is using the latest versions of its dependencies and that eslint is configured correctly.

Signed-off-by: RicardoGEsteves(Ricardo Esteves) 30535937+RicardoGEsteves@users.noreply.github.com

Refs: [#c3cd358b084a71880b03a12518cc46432eee7ea7]
Related to: [#cdbc07ff7f2d9a0688c486d757a3cb5054bbfcc9, #b580db4bf4302d6662fec5950b35db288d630ca7]